refactor(testing): unify the three attestation spec files into one module
The three attestation spec files declared the same checkpoint fields
and the checkpoint-resolution bodies in the two aggregated variants
were byte-for-byte identical (~45 duplicated lines plus 13 parallel
field declarations each).

Changes:

- one attestation_specs module replaces the three files
- a shared AttestationSpec base declares the common checkpoint fields
  and resolves target, head, and source once; the source default is a
  caller-provided checkpoint, since the aggregated paths fall back to
  the latest justified checkpoint while the single-validator gossip
  path falls back to the anchor block
- AggregatedAttestationSpec absorbs GossipAggregatedAttestationSpec:
  the two classes were field-identical, so the merged class carries
  both builders (signed gossip aggregation and invalid block-embedded
  proof); all call sites switch to the merged name
- GossipAttestationSpec keeps its single-validator semantics but
  inherits the shared resolution; its target_root_override,
  head_root_override, and source_root_override fields become
  target_root, head_root, and source_root, matching the aggregated
  naming with the same root-takes-priority cascade

Every existing test scenario resolves to identical attestation data:
the three tests using raw root overrides each set one override next
to a label whose slot already matches, so the priority change is
unobservable in emitted vectors.
