--- packages/testing/src/consensus_testing/test_types/aggregated_attestation_spec.py
@@ -1,217 +1,0 @@
-"""Lightweight aggregated attestation specification for test definitions."""
-
-from __future__ import annotations
-
-from consensus_testing.keys import XmssKeyManager
-from consensus_testing.test_types.utils import resolve_checkpoint
-from lean_spec.base import CamelModel
-from lean_spec.spec.forks import AggregationBits, Checkpoint, Slot, ValidatorIndex
-from lean_spec.spec.forks.lstar.containers import (
-    AggregatedAttestation,
-    AggregatedAttestations,
-    AttestationData,
-    Block,
-    SingleMessageAggregate,
-    State,
-)
-from lean_spec.spec.ssz import ByteList512KiB, Bytes32
-
-
-class AggregatedAttestationSpec(CamelModel):
-    """
-    Aggregated attestation specification for test definitions.
-
-    Specifies multiple validators attesting to the same data.
-    Head is derived from target.
-    Source defaults to the latest justified checkpoint unless overridden.
-    """
-
-    validator_indices: list[ValidatorIndex]
-    """The indices of validators making the attestation (required)."""
-
-    slot: Slot
-    """The slot for which the attestation is made (required)."""
-
-    target_slot: Slot
-    """The slot of the target block being attested to (required)."""
-
-    target_root_label: str | None = None
-    """
-    Label referencing a previously created block as the target.
-
-    The block must exist in the block registry with this label.
-    """
-
-    target_root: Bytes32 | None = None
-    """Optional explicit target root (bypasses label lookup)."""
-
-    head_root_label: str | None = None
-    """Optional label for the head checkpoint."""
-
-    head_root: Bytes32 | None = None
-    """Optional explicit head root."""
-
-    head_slot: Slot | None = None
-    """Optional override for the head checkpoint slot."""
-
-    source_root_label: str | None = None
-    """Optional label for the source checkpoint."""
-
-    source_root: Bytes32 | None = None
-    """Optional explicit source root."""
-
-    source_slot: Slot | None = None
-    """Optional override for the source checkpoint slot."""
-
-    valid_signature: bool = True
-    """
-    Flag whether the generated attestation signatures should be valid.
-
-    Used for testing that verification properly rejects invalid attestation signatures.
-    When False, structurally valid but cryptographically invalid signatures
-    (all zeros) will be generated for all attestations instead of proper XMSS signatures.
-
-    Defaults to True (valid signatures).
-    """
-
-    signer_ids: list[ValidatorIndex] | None = None
-    """
-    Override which validators actually sign the attestation.
-
-    When None (default), signatures are generated using the validators in validator_indices.
-    When specified, signatures are generated using these validator indices instead.
-
-    This creates a mismatch between claimed participants and actual signers.
-    Useful for testing that verification rejects attestations where valid signatures
-    don't correspond to the claimed validators.
-
-    Must have same length as validator_indices when specified.
-    """
-
-    def build_attestation_data(
-        self,
-        block_registry: dict[str, Block],
-        state: State,
-    ) -> AttestationData:
-        """
-        Build attestation data from this specification.
-
-        Args:
-            block_registry: Labeled blocks for target resolution.
-            state: State for source checkpoint lookup.
-
-        Returns:
-            Attestation data shared by all validators in the aggregation.
-
-        Raises:
-            ValueError: If no target root source is provided.
-        """
-        if self.target_root is not None:
-            target = Checkpoint(root=self.target_root, slot=self.target_slot)
-        elif self.target_root_label is not None:
-            target = resolve_checkpoint(self.target_root_label, self.target_slot, block_registry)
-        else:
-            raise ValueError("aggregated attestation spec requires a target root")
-
-        if self.head_root is not None:
-            head = Checkpoint(
-                root=self.head_root,
-                slot=self.head_slot if self.head_slot is not None else self.target_slot,
-            )
-        elif self.head_root_label is not None:
-            head = resolve_checkpoint(self.head_root_label, self.head_slot, block_registry)
-        else:
-            head = Checkpoint(
-                root=target.root,
-                slot=self.head_slot if self.head_slot is not None else target.slot,
-            )
-
-        if self.source_root is not None:
-            source = Checkpoint(
-                root=self.source_root,
-                slot=(
-                    self.source_slot
-                    if self.source_slot is not None
-                    else state.latest_justified.slot
-                ),
-            )
-        elif self.source_root_label is not None:
-            source = resolve_checkpoint(self.source_root_label, self.source_slot, block_registry)
-        else:
-            source = Checkpoint(
-                root=state.latest_justified.root,
-                slot=(
-                    self.source_slot
-                    if self.source_slot is not None
-                    else state.latest_justified.slot
-                ),
-            )
-
-        return AttestationData(
-            slot=self.slot,
-            head=head,
-            target=target,
-            source=source,
-        )
-
-    def build_invalid_proof(
-        self,
-        block_registry: dict[str, Block],
-        state: State,
-        key_manager: XmssKeyManager,
-        block: Block,
-    ) -> tuple[Block, SingleMessageAggregate]:
-        """
-        Build an invalid attestation proof and append it to the block body.
-
-        Handles two invalidity scenarios:
-
-        - Invalid signature: correct participant bitfield, zeroed-out proof bytes
-        - Signer mismatch: valid proof from wrong validators, claimed participants differ
-
-        Args:
-            block_registry: Labeled blocks for checkpoint resolution.
-            state: State for attestation data building.
-            key_manager: XMSS key manager for signing (mismatch scenario).
-            block: Current block to append the invalid attestation to.
-
-        Returns:
-            Tuple of (updated block with appended attestation, invalid proof).
-        """
-        attestation_data = self.build_attestation_data(block_registry, state)
-
-        aggregation_bits = AggregationBits.from_indices(self.validator_indices)
-        invalid_aggregated = AggregatedAttestation(
-            aggregation_bits=aggregation_bits,
-            data=attestation_data,
-        )
-
-        # Empty proof bytes flag "no real single-message aggregate here" — the caller treats
-        # any such entry as a placeholder and bypasses real binding merges.
-        placeholder = ByteList512KiB(data=b"")
-
-        if not self.valid_signature:
-            invalid_proof = SingleMessageAggregate(participants=aggregation_bits, proof=placeholder)
-        elif self.signer_ids is not None:
-            # Valid proof from wrong validators (participant mismatch).
-            valid_proof = key_manager.sign_and_aggregate(self.signer_ids, attestation_data)
-            invalid_proof = SingleMessageAggregate(
-                participants=aggregation_bits, proof=valid_proof.proof
-            )
-        else:
-            invalid_proof = SingleMessageAggregate(participants=aggregation_bits, proof=placeholder)
-
-        # Append invalid attestation to the block body.
-        block = block.model_copy(
-            update={
-                "body": block.body.model_copy(
-                    update={
-                        "attestations": AggregatedAttestations(
-                            data=[*block.body.attestations.data, invalid_aggregated]
-                        )
-                    }
-                )
-            }
-        )
-
-        return block, invalid_proof
