+++ tests/consensus/lstar/fc/test_gossip_attestation_validation.py
@@ -702,7 +702,7 @@ def test_attestation_unknown_target_block_rejected(
                     slot=Slot(2),
                     target_slot=Slot(2),
                     target_root_label="block_2",
-                    target_root_override=FAKE_ROOT,
+                    target_root=FAKE_ROOT,
                     valid_signature=False,
                 ),
                 valid=False,
@@ -742,7 +742,7 @@ def test_attestation_unknown_head_block_rejected(
                     slot=Slot(2),
                     target_slot=Slot(2),
                     target_root_label="block_2",
-                    head_root_override=FAKE_ROOT,
+                    head_root=FAKE_ROOT,
                     valid_signature=False,
                 ),
                 valid=False,
@@ -782,7 +782,7 @@ def test_attestation_unknown_source_block_rejected(
                     slot=Slot(2),
                     target_slot=Slot(2),
                     target_root_label="block_2",
-                    source_root_override=FAKE_ROOT,
+                    source_root=FAKE_ROOT,
                     valid_signature=False,
                 ),
                 valid=False,
